requirements.txt
# Основные зависимости для скальпингового торгового бота

# Асинхронный веб-фреймворк
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Асинхронный HTTP клиент
aiohttp==3.9.1

# WebSocket клиент
websockets==12.0

# Управление переменными окружения
python-dotenv==1.0.0

# Валидация данных
pydantic==2.5.3
pydantic-settings==2.1.0

# JWT аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Работа с датами и временем
python-dateutil==2.8.2

# Криптографические операции
cryptography==41.0.7

# Работа с JSON
orjson==3.9.10

# CORS поддержка
fastapi-cors==0.0.6

# Numpy для вычислений
numpy==1.26.3

# Опционально для production
# gunicorn==21.2.0
# redis==5.0.1